+++ src/types/general.ts
@@ -239,7 +239,6 @@ export interface IAdapter {
   ): Promise<ResultObjectStream>;
 
   /**
-   * @deprecated: use getPublicURL or getSignedURL
    * @param bucketName name of the bucket where the file is stored
    * @param fileName name of the file
    * @param options